Functionality to invalidate the Docker build cache for scif install.
Docker-specific

Before this commit if the contents of the scif file changed the
resulting Docker build would use the cached layer from a previous
build as the Dockerfile did not changed at all. Now the md5 hash
of the contents of the generated scif file is calculated and an
additional ARG step is added to the Dockerfile to invalidated
the Docker cache for the scif install step.