Fix bug in git: incorrect verification of commits in remote.
The git command builder was trying to verify whether a commit was
present in the remote by using `git ls-remote`. However, this
only works for commits that are the heads of remote branches.

This commit fixes that by not trying to verify commits at all,
since verification is already optional and "best effort", and
we already only selectively verify things, e.g., when both branch
and commit are specified.

To properly verify that a commit exists one would need to fetch
the whole repository and, e.g., attempt a `git checkout <commit hash>`.